requirements.txt
alpha-vantage==2.2.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
boxsdk==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cryptography==2.9.2
dnspython==1.16.0
Faker==4.1.0
idna==2.9
mongoengine==0.20.0
multidict==4.7.6
numpy==1.18.4
pandas==1.0.3
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
text-unidecode==1.3
urllib3==1.25.9
webdriver-manager==2.5.3
wrapt==1.12.1
yarl==1.4.2
